refactor: extract VideoRow into a memoized top-level component
VideoRow was defined inside App's render body, so React recreated the
component type on every render and remounted every carousel. Move it to
src/components/VideoRow.tsx as a React.memo component receiving its data
via props (shared ones spread from a single rowProps object).

Also move the pure getResolution() helper to lib/utils.ts.

App.tsx 2521 -> 2437 lines.
